requirements.txt
SpeechRecognition==3.10.0
pyttsx3==2.90
langchain-groq==0.1.5
langchain==0.1.17
python-dotenv==1.0.0
